When using Mule mock service, clientTlsContext not required
If using the mule mock service, we don’t need to set a
clientTlsContext. If we do use that things will fail since we have
not imported that certificate into the client trust store.

Assumes that the host will include "mock" (case-sensitive).
This could be a hardcode like mocksvc.mulesoft.com or a placeholder
like ${mockserver.host}.